# swagger.yaml
openapi: 3.0.3
info:
  title: FutureFarmNow API
  description: |-
    This is the API documentation for the FutureFarmNow project. It provides access to geospatial data processing and visualization, including farmland vector data, soil properties, and NDVI time-series data.
    The API is split between a Java/Scala backend for heavy-duty processing and a Python/WSGI backend for specific analytical tasks.
  version: 0.3-RC1
servers:
  - url: https://raptor.cs.ucr.edu/futurefarmnow-backend-0.3-RC1
    description: Production Server
paths:
...
  /soil/sample.json:
    post:
      summary: Request soil sample points
      description: Request a set of points to sample a field given as a polygon, based on soil properties.
      tags:
        - Soil Data
      parameters:
        - $ref: '#/components/parameters/SoilDepth'
        - name: layer
          in: query
          required: true
          description: Soil parameter(s) to consider. Can be repeated for multiple layers.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SoilLayerEnum'
          style: form
          explode: true
        - name: num_points
          in: query
          required: true
          description: The number of sample points to return.
          schema:
            type: integer
            enum: [5, 7, 10, 12]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoJSONGeometry'
      responses:
        '200':
          description: A JSON object containing the sample points and detailed statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoilSampleResponse'
    get:
      summary: Request soil sample points (GET)
      description: Request a set of points to sample a field given as a polygon, based on soil properties. GET is supported for compatibility.
      tags:
        - Soil Data
      parameters:
        - $ref: '#/components/parameters/SoilDepth'
        - name: layer
          in: query
          required: true
          description: Soil parameter(s) to consider. Can be repeated for multiple layers.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SoilLayerEnum'
          style: form
          explode: true
        - name: num_points
          in: query
          required: true
          description: The number of sample points to return.
          schema:
            type: integer
            enum: [5, 7, 10, 12]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoJSONGeometry'
      responses:
        '200':
          description: A JSON object containing the sample points and detailed statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoilSampleResponse'
...
components:
  schemas:
...
    GeoJSONGeometry:
      type: object
      properties:
        type:
          type: string
          enum: [Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon, GeometryCollection]
        coordinates:
          type: array
          items: {}
      required:
        - type
        - coordinates
...
    SoilLayerEnum:
      type: string
      enum: [alpha, bd, clay, hb, ksat, lambda, n, om, ph, sand, silt, theta_r, theta_s]
    Statistics:
      type: object
      properties:
        min:
          type: number
        max:
          type: number
        sum:
          type: number
        mean:
          type: number
        median:
          type: number
        stddev:
          type: number
        count:
          type: integer
        lowerquart:
          type: number
        upperquart:
          type: number
...
    SoilSampleResponse:
      type: object
      properties:
        query:
          $ref: '#/components/schemas/GeoJSONGeometry'
        results:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
                format: double
              y:
                type: number
                format: double
              id:
                type: integer
        statistics:
          type: object
          properties:
            layers:
              type: object
              additionalProperties:
                type: object
                properties:
                  actual:
                    $ref: '#/components/schemas/Statistics'
                  sample:
                    $ref: '#/components/schemas/Statistics'
...
  parameters:
    SoilDepth:
      name: soildepth
      in: query
      required: true
      description: "Soil depth on which to run statistics. Provided as a range 'from-to'. Accepted values for from/to are {0, 5, 15, 30, 60, 100, 200} and 'from' must be less than 'to'."
      schema:
        type: string
        pattern: '^\d+-\d+$'
      example: "0-5"
...